+++ src/lvmopstools/devices/__init__.py
@@ -5,3 +5,8 @@
 # @Date: 2024-09-11
 # @Filename: __init__.py
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
+
+from .ion import *
+from .nps import *
+from .specs import *
+from .thermistors import *